backend/pyproject.toml
[project]
name = "physical-ai-humanoid-book-backend"
version = "0.2.0"
description = "RAG Ingestion Pipeline + Agent API for Physical AI & Humanoid Robotics Textbook"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "numpy>=2.2.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.55.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]
